refactor: Phase 1 — reorganize engine into core/, tools/, skillsets/
Structural reorganization per Architecture Blueprint v1.0.
No logic changes — naming and organization only.

Directory moves:
  execution-engine/ → engine/
  supervisor/router.py → core/router.py
  supervisor/cross_team.py → core/cross_team.py
  supervisor/graph.py → core/graph_builder.py
  telegram_bot.py → core/telegram.py
  memory/tools.py → tools/memory.py
  memory/action_tools.py → tools/shell_and_web.py
  memory/service_tools.py → tools/services.py
  memory/google_tools.py → tools/google.py
  memory/obsidian_tools.py → tools/obsidian.py
  memory/unifi_tools.py → tools/unifi.py
  memory/store.py → tools/store.py

All imports updated. Docker Compose build path updated.
.venv added to .gitignore.
