+++ engine/core/__init__.py
@@ -1,0 +1,8 @@
+"""RoceOS Core — infrastructure for the AI engine.
+
+  router.py        — PA Router (classifies which skillset handles a message)
+  cross_team.py    — Multi-skillset query synthesis
+  graph_builder.py — LangGraph graph factory
+  telegram.py      — Telegram bot interface
+  llm.py           — Central model factory (cli vs api mode) [Phase 2]
+"""
