+++ engine/tools/__init__.py
@@ -1,0 +1,11 @@
+"""RoceOS Tools — capabilities that skillsets can call.
+
+Each file groups tools by domain:
+  memory.py       — remember_fact, recall_facts (persistent knowledge)
+  shell_and_web.py — SSH commands, HTTP requests, Brave search
+  services.py     — GitHub, Alpaca trading, Reddit
+  google.py       — Calendar, Gmail, Drive
+  obsidian.py     — Wiki and vault read/write
+  unifi.py        — Network device management
+  store.py        — KnowledgeStore persistence layer
+"""
